refactor package structure: move classes to a unified package and clean up imports
